+++ soh/soh/Enhancements/ExtraModes/MirroredWorld.cpp
@@ -26,7 +26,7 @@ static bool MirroredWorld_IsInDungeon(int32_t sceneNum) {
 }
 
 static void MirroredWorld_InitRandomSeed(int32_t sceneNum, uint64_t* randState) {
-    uint32_t seed =
+    uint64_t seed =
         sceneNum + (IS_RANDO ? Rando::Context::GetInstance()->GetSeed() : gSaveContext.ship.stats.fileCreatedAt);
     ShipUtils::RandInit(seed, randState);
 }
