+++ soh/soh/Enhancements/randomizer/ShuffleBeehives.cpp
@@ -112,8 +112,8 @@ void ObjComb_RandomizerUpdate(void* actor) {
     PlayState* play = gPlayState;
     combActor->unk_1B2 += 0x2EE0;
     combActor->actionFunc(combActor, play);
-    combActor->actor.shape.rot.x =
-        Math_SinS(combActor->unk_1B2) * CLAMP_MIN(combActor->unk_1B0, 0) + combActor->actor.home.rot.x;
+    combActor->actor.shape.rot.x = static_cast<s16>(Math_SinS(combActor->unk_1B2) * CLAMP_MIN(combActor->unk_1B0, 0)) +
+                                   combActor->actor.home.rot.x;
 }
 
 void RegisterShuffleBeehives() {
